Finish sketching scheduling tools, including incremental
Added a scheduleIncrementally function with the same signature as schedule, but connected to the incremental simulator.
The natural way the incremental simulator integrated with this function signature convinces me that this design is worth pursuing.
